requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-d3==3.3.11
django-geoposition==0.3.0
django-heroku==0.3.1
django-phonenumber-field==5.1.0
django-summernote==0.8.11.6
django-utils-six==2.0
gunicorn==20.1.0
heroku==0.1.4
htmllaundry==2.2
idna==2.10
lxml==4.6.3
phonenumbers==8.12.24
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==1.5
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
